backend/address_verifier/requirements.txt
Flask
Flask-Cors
python-dotenv
google-api-python-client
googlemaps
geocoder
geopy
requests
fuzzywuzzy
python-Levenshtein
pymongo
